+++ sphinxrev.cmake
@@ -13,6 +13,10 @@ function ( guess_from_git )
 		return ()
 	endif ()
 
+	# without this in some environments you can get error "detected dubious ownership in repository"
+	# `git config --global --add safe.directory '*'` in the docker image it runs in may not help. TODO: check why
+  execute_process ( COMMAND "${GIT_EXECUTABLE}" config --global --add safe.directory "${MANTICORE_SOURCE_DIR}")
+
 	# extract short has as SPH_GIT_COMMIT_ID
 	execute_process ( COMMAND "${GIT_EXECUTABLE}" log -1 --format=%h
 			WORKING_DIRECTORY "${MANTICORE_SOURCE_DIR}"
@@ -22,19 +26,17 @@ function ( guess_from_git )
 			OUTPUT_STRIP_TRAILING_WHITESPACE )
 	set ( SPH_GIT_COMMIT_ID "${SPH_GIT_COMMIT_ID}" PARENT_SCOPE )
 
-	# extract timestamp and make number YYMMDD from it
-	# it would be --date=format:%y%m%d, but old git on centos doesn't understand it
-	execute_process ( COMMAND "${GIT_EXECUTABLE}" log -1 --date=short --format=%cd
+	# extract timestamp and make number YYMMDDHH from it
+	execute_process ( COMMAND "${GIT_EXECUTABLE}" log -1 --date=format:"%y%m%d%H" --format=%cd
 			WORKING_DIRECTORY "${MANTICORE_SOURCE_DIR}"
 			RESULT_VARIABLE res
 			OUTPUT_VARIABLE GIT_TIMESTAMP_ID
-			ERROR_QUIET
 			OUTPUT_STRIP_TRAILING_WHITESPACE )
-	string ( REPLACE "-" "" GIT_TIMESTAMP_ID "${GIT_TIMESTAMP_ID}" )
-	string ( SUBSTRING "${GIT_TIMESTAMP_ID}" 2 -1 GIT_TIMESTAMP_ID )
-	set ( GIT_TIMESTAMP_ID "${GIT_TIMESTAMP_ID}" PARENT_SCOPE )
 
-	# timestamp for reproducable packages
+	string ( SUBSTRING "${GIT_TIMESTAMP_ID}" 1 8 GIT_TIMESTAMP_ID )
+	set ( GIT_TIMESTAMP_ID ${GIT_TIMESTAMP_ID} PARENT_SCOPE )
+
+	# timestamp for reproducible packages
 	execute_process ( COMMAND "${GIT_EXECUTABLE}" log -1 --pretty=%ct
 			WORKING_DIRECTORY "${MANTICORE_SOURCE_DIR}"
 			RESULT_VARIABLE res