+++ src/schema/schema.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2017-2023, Manticore Software LTD (https://manticoresearch.com)
+// Copyright (c) 2017-2024, Manticore Software LTD (https://manticoresearch.com)
 // Copyright (c) 2001-2016, Andrew Aksyonoff
 // Copyright (c) 2008-2016, Sphinx Technologies Inc
 // All rights reserved
@@ -355,11 +355,11 @@ bool CSphSchema::IsReserved ( const char* szToken )
 {
 	static const char * dReserved[] =
 	{
-		"AND", "AS", "BY", "COLUMNARSCAN", "DISTINCT", "DIV", "DOCIDINDEX", "EXPLAIN",
-		"FACET", "FALSE", "FORCE", "FROM", "IGNORE", "IN", "INDEXES", "IS", "KNN", "LIMIT",
-		"MOD", "NOT", "NO_COLUMNARSCAN", "NO_DOCIDINDEX", "NO_SECONDARYINDEX", "NULL",
-		"OFFSET", "OR", "ORDER", "REGEX", "RELOAD", "SECONDARYINDEX", "SELECT", "SYSFILTERS",
-		"TRUE", NULL
+		"AND", "AS", "BY", "COLUMNARSCAN", "DATE_ADD", "DATE_SUB", "DAY", "DISTINCT", "DIV", "DOCIDINDEX", "EXPLAIN",
+		"FACET", "FALSE", "FORCE", "FROM", "HOUR", "IGNORE", "IN", "INDEXES", "INTERVAL", "IS", "KNN", "LIMIT",
+		"MINUTE", "MOD", "MONTH", "NOT", "NO_COLUMNARSCAN", "NO_DOCIDINDEX", "NO_SECONDARYINDEX", "NULL",
+		"OFFSET", "OR", "ORDER", "QUARTER", "REGEX", "RELOAD", "SECOND", "SECONDARYINDEX", "SELECT", "SYSFILTERS",
+		"TRUE", "WEEK", "YEAR", NULL
 	};
 
 	const char** p = dReserved;